# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "daisy-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.8.0"
  },
  "paths": {
...
    "/sessions": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "List of available sessions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session_history",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Annotations": {
        "type": "object",
        "properties": {
          "audience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "priority": {
            "type": "number"
          }
        }
      },
      "Author": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ContextLengthExceeded": {
        "type": "object",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
...
      "ExtensionData": {
        "type": "object",
        "description": "Extension data containing all extension states\nKeys are in format \"extension_name.version\" (e.g., \"todo.v0\")",
        "additionalProperties": {}
      },
...
      "FrontendToolRequest": {
        "type": "object",
        "required": [
          "id",
          "toolCall"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "toolCall": {
            "type": "object"
          }
        }
      },
...
      "ImageContent": {
        "type": "object",
        "required": [
          "data",
          "mimeType"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "annotations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Annotations"
              },
              {
                "type": "object"
              }
            ]
          },
          "data": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "MessageContent": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/TextContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "text"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolResponse"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolResponse"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolConfirmationRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolConfirmationRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FrontendToolRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "frontendToolRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ThinkingContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "thinking"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RedactedThinkingContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "redactedThinking"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContextLengthExceeded"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "contextLengthExceeded"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SummarizationRequested"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "summarizationRequested"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "description": "Content passed inside a message, which can be both simple content and tool content",
        "discriminator": {
          "propertyName": "type"
        }
      },
      "MessageMetadata": {
        "type": "object",
        "description": "Metadata for message visibility",
        "properties": {
          "agentVisible": {
            "type": "boolean",
            "description": "Whether the message should be included in the agent's context window"
          },
          "userVisible": {
            "type": "boolean",
            "description": "Whether the message should be visible to the user in the UI"
          }
        }
      },
...
      "Recipe": {
        "type": "object",
        "description": "A Recipe represents a personalized, user-generated agent configuration that defines\nspecific behaviors and capabilities within the Daisy system.\n\n# Fields\n\n## Required Fields\n* `version` - Semantic version of the Recipe file format (defaults to \"1.0.0\")\n* `title` - Short, descriptive name of the Recipe\n* `description` - Detailed description explaining the Recipe's purpose and functionality\n* `Instructions` - Instructions that defines the Recipe's behavior\n\n## Optional Fields\n* `prompt` - the initial prompt to the session to start with\n* `extensions` - List of extension configurations required by the Recipe\n* `context` - Supplementary context information for the Recipe\n* `activities` - Activity labels that appear when loading the Recipe\n* `author` - Information about the Recipe's creator and metadata\n* `parameters` - Additional parameters for the Recipe\n* `response` - Response configuration including JSON schema validation\n* `retry` - Retry configuration for automated validation and recovery\n# Example\n\n\nuse daisy::recipe::Recipe;\n\n// Using the builder pattern\nlet recipe = Recipe::builder()\n.title(\"Example Agent\")\n.description(\"An example Recipe configuration\")\n.instructions(\"Act as a helpful assistant\")\n.build()\n.expect(\"Missing required fields\");\n\n// Or using struct initialization\nlet recipe = Recipe {\nversion: \"1.0.0\".to_string(),\ntitle: \"Example Agent\".to_string(),\ndescription: \"An example Recipe configuration\".to_string(),\ninstructions: Some(\"Act as a helpful assistant\".to_string()),\nprompt: None,\nextensions: None,\ncontext: None,\nactivities: None,\nauthor: None,\nsettings: None,\nparameters: None,\nresponse: None,\nsub_recipes: None,\nretry: None,\n};\n",
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ],
            "nullable": true
          },
          "context": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionConfig"
            },
            "nullable": true
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeParameter"
            },
            "nullable": true
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Response"
              }
            ],
            "nullable": true
          },
          "retry": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RetryConfig"
              }
            ],
            "nullable": true
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ],
            "nullable": true
          },
          "sub_recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubRecipe"
            },
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "RedactedThinkingContent": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "string"
          }
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "json_schema": {
            "nullable": true
          }
        }
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "SessionHistoryResponse": {
        "type": "object",
        "required": [
          "sessionId",
          "metadata",
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "List of messages in the session conversation"
          },
          "metadata": {
            "$ref": "#/components/schemas/SessionMetadata"
          },
          "sessionId": {
            "type": "string",
            "description": "Unique identifier for the session"
          }
        }
      },
      "SessionInfo": {
        "type": "object",
        "required": [
          "id",
          "path",
          "modified",
          "metadata"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/SessionMetadata"
          },
          "modified": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "SessionListResponse": {
        "type": "object",
        "required": [
          "sessions"
        ],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionInfo"
            },
            "description": "List of available session information objects"
          }
        }
      },
      "SessionMetadata": {
        "type": "object",
        "description": "Metadata for a session, stored as the first line in the session file",
        "required": [
          "working_dir",
          "description",
          "message_count"
        ],
        "properties": {
          "accumulated_input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of input tokens used in the session. Accumulated across all messages.",
            "nullable": true
          },
          "accumulated_output_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of output tokens used in the session. Accumulated across all messages.",
            "nullable": true
          },
          "accumulated_total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used in the session. Accumulated across all messages (useful for tracking cost over an entire session).",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "A short description of the session, typically 3 words or less"
          },
          "extension_data": {
            "$ref": "#/components/schemas/ExtensionData"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of input tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "description": "Number of messages in the session",
            "minimum": 0
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of output tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "description": "ID of the schedule that triggered this session, if any",
            "nullable": true
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "working_dir": {
            "type": "string",
            "description": "Working directory for the session",
            "example": "/home/user/sessions/session1"
          }
        }
      },
...
      "SummarizationRequested": {
        "type": "object",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
      "TextContent": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "_meta": {
            "type": "object",
            "additionalProperties": true
          },
          "annotations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Annotations"
              },
              {
                "type": "object"
              }
            ]
          },
          "text": {
            "type": "string"
          }
        }
      },
      "ThinkingContent": {
        "type": "object",
        "required": [
          "thinking",
          "signature"
        ],
        "properties": {
          "signature": {
            "type": "string"
          },
          "thinking": {
            "type": "string"
          }
        }
      },
...
      "ToolConfirmationRequest": {
        "type": "object",
        "required": [
          "id",
          "toolName",
          "arguments"
        ],
        "properties": {
          "arguments": {},
          "id": {
            "type": "string"
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "toolName": {
            "type": "string"
          }
        }
      },
...
      "ToolRequest": {
        "type": "object",
        "required": [
          "id",
          "toolCall"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "toolCall": {
            "type": "object"
          }
        }
      },
      "ToolResponse": {
        "type": "object",
        "required": [
          "id",
          "toolResult"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "toolResult": {
            "type": "object"
          }
        }
      },
...
    }
  }
}